Modernize EcalLaserCorrectionService
- move header to source
- migrate to EventSetup consumes
- make it shorter
