+++ CalibCalorimetry/EcalLaserCorrection/plugins/EcalLaserCorrectionService.cc
@@ -5,13 +5,47 @@
 // system include files
 #include <iostream>
 #include <fstream>
+#include <memory>
 
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/Framework/interface/ESProductHost.h"
+#include "FWCore/Utilities/interface/ReusableObjectHolder.h"
 #include "FWCore/Framework/interface/ESHandle.h"
 
 #include "CalibCalorimetry/EcalLaserCorrection/interface/EcalLaserDbService.h"
 #include "CalibCalorimetry/EcalLaserCorrection/interface/EcalLaserDbRecord.h"
 
-#include "CalibCalorimetry/EcalLaserCorrection/plugins/EcalLaserCorrectionService.h"
+#include "CondFormats/DataRecord/interface/EcalLaserAlphasRcd.h"
+#include "CondFormats/DataRecord/interface/EcalLaserAPDPNRatiosRefRcd.h"
+#include "CondFormats/DataRecord/interface/EcalLaserAPDPNRatiosRcd.h"
+#include "CondFormats/DataRecord/interface/EcalLinearCorrectionsRcd.h"
+
+class EcalLaserCorrectionService : public edm::ESProducer {
+public:
+  EcalLaserCorrectionService(const edm::ParameterSet&);
+  ~EcalLaserCorrectionService() override;
+
+  std::shared_ptr<EcalLaserDbService> produce(const EcalLaserDbRecord&);
+
+private:
+  using HostType = edm::ESProductHost<EcalLaserDbService,
+                                      EcalLaserAlphasRcd,
+                                      EcalLaserAPDPNRatiosRefRcd,
+                                      EcalLaserAPDPNRatiosRcd,
+                                      EcalLinearCorrectionsRcd>;
+
+  // ----------member data ---------------------------
+  edm::ReusableObjectHolder<HostType> holder_;
+
+  edm::ESGetToken<EcalLaserAlphas, EcalLaserAlphasRcd> alphaToken_;
+  edm::ESGetToken<EcalLaserAPDPNRatiosRef, EcalLaserAPDPNRatiosRefRcd> apdpnRefToken_;
+  edm::ESGetToken<EcalLaserAPDPNRatios, EcalLaserAPDPNRatiosRcd> apdpnToken_;
+  edm::ESGetToken<EcalLinearCorrections, EcalLinearCorrectionsRcd> linearToken_;
+
+  //  std::vector<std::string> mDumpRequest;
+  //  std::ostream* mDumpStream;
+};
 
 EcalLaserCorrectionService::EcalLaserCorrectionService(const edm::ParameterSet& fConfig)
     : ESProducer()
@@ -22,7 +56,11 @@ EcalLaserCorrectionService::EcalLaserCorrectionService(const edm::ParameterSet& 
   // data is being produced
   //  setWhatProduced (this, (dependsOn (&EcalLaserCorrectionService::apdpnCallback)));
 
-  setWhatProduced(this);
+  setWhatProduced(this)
+      .setConsumes(alphaToken_)
+      .setConsumes(apdpnRefToken_)
+      .setConsumes(apdpnToken_)
+      .setConsumes(linearToken_);
 
   //now do what ever other initialization is needed
 
@@ -47,40 +85,19 @@ EcalLaserCorrectionService::~EcalLaserCorrectionService() {
 std::shared_ptr<EcalLaserDbService> EcalLaserCorrectionService::produce(const EcalLaserDbRecord& record) {
   auto host = holder_.makeOrGet([]() { return new HostType; });
 
-  host->ifRecordChanges<EcalLinearCorrectionsRcd>(record,
-                                                  [this, h = host.get()](auto const& rec) { setupLinear(rec, *h); });
+  host->ifRecordChanges<EcalLinearCorrectionsRcd>(
+      record, [this, h = host.get()](auto const& rec) { h->setLinearCorrectionsData(&rec.get(linearToken_)); });
 
-  host->ifRecordChanges<EcalLaserAPDPNRatiosRcd>(record,
-                                                 [this, h = host.get()](auto const& rec) { setupApdpn(rec, *h); });
+  host->ifRecordChanges<EcalLaserAPDPNRatiosRcd>(
+      record, [this, h = host.get()](auto const& rec) { h->setAPDPNData(&rec.get(apdpnToken_)); });
 
   host->ifRecordChanges<EcalLaserAPDPNRatiosRefRcd>(
-      record, [this, h = host.get()](auto const& rec) { setupApdpnRef(rec, *h); });
+      record, [this, h = host.get()](auto const& rec) { h->setAPDPNRefData(&rec.get(apdpnRefToken_)); });
 
-  host->ifRecordChanges<EcalLaserAlphasRcd>(record, [this, h = host.get()](auto const& rec) { setupAlpha(rec, *h); });
+  host->ifRecordChanges<EcalLaserAlphasRcd>(
+      record, [this, h = host.get()](auto const& rec) { h->setAlphaData(&rec.get(alphaToken_)); });
 
   return host;  // automatically converts to std::shared_ptr<EcalLaserDbService>
 }
 
-void EcalLaserCorrectionService::setupAlpha(const EcalLaserAlphasRcd& fRecord, EcalLaserDbService& service) {
-  edm::ESHandle<EcalLaserAlphas> item;
-  fRecord.get(item);
-  service.setAlphaData(item.product());
-}
-
-void EcalLaserCorrectionService::setupApdpnRef(const EcalLaserAPDPNRatiosRefRcd& fRecord, EcalLaserDbService& service) {
-  edm::ESHandle<EcalLaserAPDPNRatiosRef> item;
-  fRecord.get(item);
-  service.setAPDPNRefData(item.product());
-}
-
-void EcalLaserCorrectionService::setupApdpn(const EcalLaserAPDPNRatiosRcd& fRecord, EcalLaserDbService& service) {
-  edm::ESHandle<EcalLaserAPDPNRatios> item;
-  fRecord.get(item);
-  service.setAPDPNData(item.product());
-}
-
-void EcalLaserCorrectionService::setupLinear(const EcalLinearCorrectionsRcd& fRecord, EcalLaserDbService& service) {
-  edm::ESHandle<EcalLinearCorrections> item;
-  fRecord.get(item);
-  service.setLinearCorrectionsData(item.product());
-}
+DEFINE_FWK_EVENTSETUP_MODULE(EcalLaserCorrectionService);
