--- CalibCalorimetry/EcalLaserCorrection/plugins/SealModule.cc
@@ -1,5 +1,0 @@
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-
-#include "CalibCalorimetry/EcalLaserCorrection/plugins/EcalLaserCorrectionService.h"
-
-DEFINE_FWK_EVENTSETUP_MODULE(EcalLaserCorrectionService);
